moved all csr classes and config into new module
