+++ .idea/compiler.xml
@@ -23,18 +23,20 @@
       <module name="google-account_test" target="1.8" />
       <module name="google-cloud-core_main" target="1.8" />
       <module name="google-cloud-core_test" target="1.8" />
+      <module name="google-cloud-repos_main" target="1.8" />
+      <module name="google-cloud-repos_test" target="1.8" />
       <module name="google-cloud-storage_main" target="1.8" />
       <module name="google-cloud-storage_test" target="1.8" />
       <module name="google-cloud-tools-plugin-ultimate_main" target="1.8" />
       <module name="google-cloud-tools-plugin-ultimate_test" target="1.8" />
       <module name="google-cloud-tools-plugin_main" target="1.8" />
       <module name="google-cloud-tools-plugin_test" target="1.8" />
-      <module name="stackdriver-debugger_main" target="1.8" />
-      <module name="stackdriver-debugger_test" target="1.8" />
       <module name="java-ultimate_main" target="1.8" />
       <module name="java-ultimate_test" target="1.8" />
       <module name="java_main" target="1.8" />
       <module name="java_test" target="1.8" />
+      <module name="stackdriver-debugger_main" target="1.8" />
+      <module name="stackdriver-debugger_test" target="1.8" />
       <module name="ultimate_main" target="1.8" />
       <module name="ultimate_test" target="1.8" />
     </bytecodeTargetLevel>
