+++ google-cloud-tools-plugin/google-cloud-repos/build.gradle
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// Use IntelliJ src structure for now
+sourceSets.main.java.srcDirs = ['src']
+// 'src' is included in resources to export .form files to upstream modules.
+sourceSets.main.resources.srcDirs = ['resources', 'src']
+sourceSets.test.java.srcDirs = ['testSrc']
+sourceSets.test.resources.srcDirs = ['testResources']
+
+intellij {
+    plugins 'git4idea'
+}
+
+dependencies {
+    compile(project(':google-cloud-tools-plugin:google-cloud-core'))
+    compile files('lib/google-api-services-source.jar')
+
+    compileOnly 'com.google.auto.value:auto-value:1.4.1'
+    apt         'com.google.auto.value:auto-value:1.4.1'
+
+    testCompile files("${System.properties['java.home']}/../lib/tools.jar")
+    testCompile(project(':common-test-lib'))
+}
+
+configurations {
+    // these are already obtained from google-api-java-client-min-repackaged
+    compile.exclude group: 'com.google.api-client', module: 'google-api-client'
+    compile.exclude group: 'com.google.http-client', module: 'google-http-client'
+}
