+++ google-cloud-tools-plugin/google-cloud-repos/resources/META-INF/google-cloud-repos.xml
@@ -1,0 +1,46 @@
+<!--
+  ~ Copyright 2018 Google Inc. All Rights Reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<idea-plugin>
+    <depends>com.intellij.modules.vcs</depends>
+    <depends>Git4Idea</depends>
+    <extensions defaultExtensionNs="com.intellij">
+        <checkoutProvider implementation="com.google.cloud.tools.intellij.csr.GcpCheckoutProvider"/>
+        <vcsPopupProvider
+                implementation="com.google.cloud.tools.intellij.csr.GcpQuickListContentProvider"/>
+        <applicationService
+                serviceImplementation="com.google.cloud.tools.intellij.csr.CloudRepositoryService"/>
+    </extensions>
+
+    <extensions defaultExtensionNs="Git4Idea">
+        <GitHttpAuthDataProvider
+                implementation="com.google.cloud.tools.intellij.csr.GcpHttpAuthDataProvider"/>
+    </extensions>
+
+    <actions>
+        <action id="GoogleCloudTools.UploadSourceToGCP"
+                class="com.google.cloud.tools.intellij.csr.SetupCloudRepositoryAction">
+            <add-to-group group-id="Vcs.Import"/>
+        </action>
+
+        <group id="GoogleCloudTools.csr">
+            <separator/>
+            <reference ref="GoogleCloudTools.UploadSourceToGCP"/>
+            <separator/>
+            <add-to-group group-id="GoogleCloudTools"/>
+        </group>
+    </actions>
+</idea-plugin>
