+++ google-cloud-tools-plugin/google-cloud-repos/resources/messages/CloudReposBundle.properties
@@ -65,3 +65,8 @@ cloud.repository.selector.placeholder.text=Enter a cloud repository ID...
 cloud.repository.selector.missing.project.error=Select a Google Cloud project
 cloud.repository.selector.create.button=Create a new cloud repository
 cloud.repository.selector.create.url.error=Failed to load GCP create repository URL
+
+httpauthprovider.chooselogin=Choose a Google Login to perform this action.
+uploadsourceaction.addfiles=Adding files to git...
+uploadsourceaction.performingcommit=Performing commit...
+uploadsourceaction.addfilestitle=Add Files For Initial Commit
