+++ google-cloud-tools-plugin/google-cloud-repos/src/com/google/cloud/tools/intellij/csr/GcpHttpAuthDataProvider.java
@@ -20,7 +20,6 @@ import com.google.api.client.repackaged.com.google.common.base.Strings;
 import com.google.cloud.tools.intellij.login.CredentialedUser;
 import com.google.cloud.tools.intellij.login.Services;
 import com.google.cloud.tools.intellij.resources.SelectUserDialog;
-import com.google.cloud.tools.intellij.util.CloudReposMessageBundle;
 import com.intellij.ide.DataManager;
 import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
@@ -33,7 +32,7 @@ import com.intellij.openapi.wm.ex.WindowManagerEx;
 import com.intellij.util.AuthData;
 import git4idea.DialogManager;
 import git4idea.remote.GitHttpAuthDataProvider;
-import java.awt.Window;
+import java.awt.*;
 import java.io.IOException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
