+++ google-cloud-tools-plugin/google-cloud-repos/src/com/google/cloud/tools/intellij/csr/GcpQuickListContentProvider.java
