+++ google-cloud-tools-plugin/google-cloud-repos/testSrc/com/google/cloud/tools/intellij/csr/GcpHttpAuthProviderTest.java.excluded
