+++ google-cloud-tools-plugin/resources/META-INF/plugin.xml
@@ -64,16 +64,12 @@
 
 
   <extensions defaultExtensionNs="com.intellij">
-     <checkoutProvider implementation="com.google.cloud.tools.intellij.csr.GcpCheckoutProvider"/>
-    <vcsPopupProvider implementation="com.google.cloud.tools.intellij.csr.GcpQuickListContentProvider"/>
-
 
     <applicationService serviceInterface="com.google.cloud.tools.intellij.service.ApplicationPluginInfoService"
                         serviceImplementation="com.google.cloud.tools.intellij.service.DefaultApplicationPluginInfoService"/>
 
     <applicationService serviceImplementation="com.google.cloud.tools.intellij.apis.CloudLibrariesService"/>
     <applicationService serviceImplementation="com.google.cloud.tools.intellij.apis.CloudApiMavenService"/>
-    <applicationService serviceImplementation="com.google.cloud.tools.intellij.csr.CloudRepositoryService"/>
 
     <localInspection language="XML" shortName="DependencyVersionWithBom" bundle="messages.CloudToolsBundle"
                      hasStaticDescription="true" key="cloud.libraries.version.with.bom.inspection.display.name"
@@ -87,10 +83,6 @@
 
   </extensions>
 
-  <extensions defaultExtensionNs="Git4Idea">
-    <GitHttpAuthDataProvider implementation="com.google.cloud.tools.intellij.csr.GcpHttpAuthDataProvider" />
-  </extensions>
-
   <extensions defaultExtensionNs="com.google.gct.core">
     <googleLoginMessageExtender
             implementation="com.google.cloud.tools.intellij.login.IntelliJGoogleLoginMessageExtender"/>
@@ -101,18 +93,11 @@
   </extensions>
 
   <actions>
-    <action id="GoogleCloudTools.UploadSourceToGCP"
-            class="com.google.cloud.tools.intellij.csr.SetupCloudRepositoryAction">
-      <add-to-group group-id="Vcs.Import"/>
-    </action>
 
     <group id="GoogleCloudTools.misc">
       <separator/>
       <action id="GoogleCloudTools.CloudDebugger"
               class="com.google.cloud.tools.intellij.stackdriver.debugger.CloudDebuggerToolsMenuAction"/>
-      <separator/>
-      <reference ref="GoogleCloudTools.UploadSourceToGCP"/>
-      <separator/>
       <action id="GoogleCloudTools.CloudStorage"
               class="com.google.cloud.tools.intellij.gcs.GcsToolWindowAction"/>
       <separator/>
