+++ settings.gradle
@@ -17,6 +17,7 @@ include 'google-cloud-tools-plugin'
 include 'common-test-lib'
 include 'google-cloud-tools-plugin:google-cloud-core'
 include 'google-cloud-tools-plugin:google-cloud-storage'
+include 'google-cloud-tools-plugin:google-cloud-repos'
 include 'google-cloud-tools-plugin:stackdriver-debugger'
 include 'google-cloud-tools-plugin:app-engine:java'
 include 'google-cloud-tools-plugin:app-engine:java:ultimate'
